new_agent/requirements.txt
openai>=1.52.0
python-dotenv>=1.0.1
rich>=13.9.2
